+++ R/globals.R
@@ -1,5 +1,6 @@
 utils::globalVariables(c("RMSE", "adjR", "control", "controlID",
-                         "control_coefs", "feols", "fixed", "index", "p",
-                         "ribbon.group", "se", "sig.level", "statistic",
-                         "term", "value", "."))
+                         "control_coefs", "estimate", "estimate_FE", "feols",
+                         "fixed", "index", "is_fe", "lower", "p",
+                         "ribbon.group", "se", "se_type", "sig", "sig.level",
+                         "statistic", "term", "upper", "value", "."))